// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ROLE = {
  ADMIN: 'ADMIN',
  USER: 'USER'
} as const

export type ROLE = (typeof ROLE)[keyof typeof ROLE]


export const WeekDay = {
  Monday: 'Monday',
  Tuesday: 'Tuesday',
  Wednesday: 'Wednesday',
  Thursday: 'Thursday',
  Friday: 'Friday',
  Saturday: 'Saturday',
  Sunday: 'Sunday'
} as const

export type WeekDay = (typeof WeekDay)[keyof typeof WeekDay]


export const TaskStatus = {
  PENDING: 'PENDING',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  SKIPPED: 'SKIPPED'
} as const

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus]